requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
werkzeug==3.0.4
qrcode[pil]==7.4.2
pillow>=12.1.0
gunicorn==22.0.0
psycopg2-binary==2.9.11
cloudinary==1.42.0